게시물 삭제 시 권한 체크

+++ src/main/java/com/pjs/exam/demo/controller/UsrArticleController.java
@@ -12,6 +12,8 @@ import com.pjs.exam.demo.util.Ut;
 import com.pjs.exam.demo.vo.Article;
 import com.pjs.exam.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 	@Autowired
@@ -20,14 +22,25 @@ public class UsrArticleController {
 	//액션 메서드 시작
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData<Article> doAdd(String title, String body) {
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId")!=null){
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		if(isLogined == false) {
+			return ResultData.from("F-4", "로그인 후 이용해주세요.");
+		}
+		
 		if(Ut.empty(title)) {
 			return ResultData.from("F-1", "title(을)를 입력해주세요");
 		}
 		if(Ut.empty(body)) {
 			return ResultData.from("F-2", "body(을)를 입력해주세요");
 		}
-		ResultData<Integer> writeArticleRd = articleService.writeArticle(title,body);
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title,body);
 		int id = (int)writeArticleRd.getData1();
 		
 		Article article = articleService.getArticle(id);
@@ -55,13 +68,29 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId")!=null){
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		if(isLogined == false) {
+			return ResultData.from("F-4", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticle(id);
+		if(article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", "로그인 후 이용해주세요.");
+		}
+		
 		if(article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 		}
 		articleService.deleteArticle(id);
-		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제하였습니다.", id),id);
+		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제하였습니다.", id), id);
 	}
 	
 	@RequestMapping("/usr/article/doModify")
